AllerMatch website/backend/CommandTool/requirements.txt
biopython
exrex
fpdf
pandas
bs4
lxml
click